Update link opening style (#4262)
* updated opening styles by editing link_contoller.js and relevant <a> tags, visually-hidden tag still appearing in console

* commented out ExternalLinks in HTMLResponseTransformer to test whether this removes visually-hidden tags from external links

* created new-tab class to be used for links that need to open in new tab / created test, deleted external links class

* modified link controller and test, adapted NewTabLinks spec file, added new-tab class to chat component

* removed new tab attributes from new footer

* added a condition to NewTabLinks to handle links styled as buttons

* added a test for links with class of new-tab and button

* modified chat component test to reflect new wording of link: Open chat in new tab

* updated form links for get an adviser to open in new tab

* updated links to open in a new tab on get a teacher training adviser - which country do you live in page

* removed new-class tab on no degree summary page

* added aria labels to social media links in footer, for testing
