+++ app/views/teacher_training_adviser/steps/_no_degree.html.erb
@@ -5,6 +5,6 @@
 
     <p>You need a bachelor’s degree to train to teach in primary, secondary and special schools in England.</p>
     <p>If you do not already have one, you can <%= link_to("train to be a qualified teacher as part of your bachelor’s degree", "/train-to-be-a-teacher/if-you-dont-have-a-degree") %>.</p>
-    <p>You can also <%= link_to("teach in further education", "https://www.teach-in-further-education.campaign.gov.uk", target: "_blank", rel: :noopener) %> without a degree.</p>
+    <p>You can also <%= link_to("teach in further education", "https://www.teach-in-further-education.campaign.gov.uk") %> without a degree.</p>
     <p><%= link_to("Find out more about the qualifications you need to teach", "/train-to-be-a-teacher/qualifications-you-need-to-teach") %>.</p>
 </div>
